feat(axum): implement feature gates for default pages
Reorganize the codebase to support optional UI components through feature flags:
- Add `default-pages` feature (enabled by default)
- Split core user functionality from UI components
- Move user management to dedicated module
- Improve function visibility by making handlers private where possible
- Conditionally merge routers based on feature flags

This change allows consumers to disable the default UI pages when not needed,
making the library more flexible while maintaining backward compatibility.
The default behavior remains unchanged, but users can now opt out of the
UI components by using `--no-default-features` during build.
