+++ oauth2_passkey_axum/src/lib.rs
@@ -1,11 +1,13 @@
 mod config;
+#[cfg(feature = "default-pages")]
+mod default_pages;
 mod error;
 mod middleware;
 mod oauth2;
-mod pages;
 mod passkey;
 mod router;
 mod session;
+mod user;
 
 pub use config::{O2P_REDIRECT_ANON, O2P_REDIRECT_USER};
 pub use middleware::{
