--- oauth2_passkey_axum/src/pages/mod.rs
@@ -1,6 +1,0 @@
-mod login;
-mod logout;
-mod router;
-mod user;
-
-pub(crate) use router::router as user_router;
